Apply PR feedbacks.
